minimal NSUInteger fix
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gworkspace/trunk@39702 72102866-910b-0410-8b05-ffd578937521

+++ Operation/FileOpInfo.m
@@ -145,7 +145,7 @@ static NSString *nibName = @"FileOperationWin";
 	  NSString *base = [newname stringByDeletingPathExtension];        
 	  NSString *ntmp;
 	  NSString *destpath;        
-	  int count = 1;
+	  NSUInteger count = 1;
 	      
 	  while (1)
 	    {
